+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add tooltips by Tippy.js
 - Add tooltips for Bar-Chart
+- Add tooltips for Line-Chart
 
 ## [0.2.3] - 2019-10-17